% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogramr}
\alias{histogramr}
\title{Irregular histograms}
\usage{
histogramr(x, breaks, method = c("greedy", "exact"),
  weights = c("unequal", "equal"), type = c("KL", "L2"), plot = TRUE,
  control, ...)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired. \code{NA}
values are automatically removed. \code{x} must be contained contained
in \eqn{[0, 1]}.}

\item{breaks}{a single number giving the number of cells for the histogram.}

\item{method}{the method to be used. See ‘Details’. Can be abbreviated.}

\item{weights}{the weights to be used. See ‘Details’. Can be abbreviated.}

\item{type}{the type of discrepancy to minimize. See ‘Details’.}

\item{plot}{logical. If \code{TRUE} (default), a histogram is plotted,
otherwise a list of breaks and counts is returned.}

\item{control}{a list of control parameters. See ‘Details’.}

\item{...}{further arguments and graphical parameters passed to
\code{plot.histogramr}.}
}
\description{
The function \code{histogramr} computes an irregular histogram of the given
data values between 0 and 1. If \code{plot = TRUE}, the resulting object of
class \code{histogramr} is plotted by \code{plot.histogramr}, before
it is returned.
}
\details{
The \code{method} option can be either of \code{"greedy"} or \code{"exact"}. The
   \code{"exact"} algorithm uses dynamic programming to find the true maximum
   likelihood estimator (if \code{type = "KL"}) or the true L2-minimzer
    (if \code{type = "L2"}). This method suffers from two drawbacks.
   It is considerable slower thant the
   greedy algorithm and tends to overfit. Choosing \code{method = "greedy"}
   is recommended. This runs a greedy coordinate search, which is much faster
   than the exact algorithm and is less prone to overfitting. When
   \code{greedy} is used, the options \code{init} and \code{modulator} can
   be passed to \code{control}. \code{init} specifies the inital break points
   in the coordinate search, while \code{modulator} limits the number of
   times the algorithms runs.

If \code{weights} is \code{"uequal"} an ordinary irregular histogram is
   estimated, but if \code{weights} is \code{"equal"} an histogram with
   equal areas for every bin is estimated instead, see
   (Denby and Mallows, 2009). The option \code{"uequal"} is recommended.

For maximum likelihood estimation choose \code{type} equal to \code{"KL"}.
   Choose \code{L2} for minimization of the L2 discrepancy. The \code{"L2"}
   option is slightly faster but \code{"KL"} gives slightly better results.
   The option \code{"KL"} is recommended.

The \code{control} argument is a list that can supply any of the following
   components:
   \itemize{
    \item{\code{eps} The minimal distance between two break points. Defaults
      to \eqn{n^{-1/2}}}
    \item{\code{init} Only for \code{method = "greedy"}. Specifies the
      inital break points in the coordinate search.}
    \item{\code{modulator} Only for \code{method = "greedy"}. Limits the
       number of times the algorithm runs.}
   }
}
\references{
Denby, L., & Mallows, C. (2009). Variations on the histogram. Journal of
Computational and Graphical Statistics, 18(1), 21-31.
}
