% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_comp.R
\name{nc_comp}
\alias{nc_comp}
\alias{nc_comp.default}
\alias{print.nc_comp}
\alias{ggplot.nc_comp}
\alias{summary.nc_comp}
\title{Comparison of methods to calculate the number of classes}
\usage{
nc_comp(x, density = FALSE, xlab = NULL, ...)

\method{nc_comp}{default}(x, density = FALSE, xlab = NULL, ...)

\method{print}{nc_comp}(x, ...)

\method{ggplot}{nc_comp}(x, ...)

\method{summary}{nc_comp}(x, ...)
}
\arguments{
\item{x}{a numerical vector.}

\item{density}{logical; if \code{TRUE}, the histogram graphic is a representation of frequencies, the counts component of the result; if \code{FALSE}, probability densities, component density, are plotted (so that the histogram has a total area of one).}

\item{xlab}{a character string, X axis label. By default it gets from object vector name with \code{deparse(substitute(x))}.}
}
\value{
A list of values.
}
\description{
Compare the methods for calculation of the number of classes from a numerical random variable.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default output from \code{nc_comp} function.

\item \code{nc_comp}: Print output from \code{nc_comp} function.

\item \code{nc_comp}: ggplot2 plot output from \code{nc_comp} function.

\item \code{nc_comp}: Summary output from \code{nc_comp} function.
}}

\seealso{
\code{\link{hist}}, \code{\link{geom_histogram}}, \code{\link{geom_freqpoly}}.
}
