% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.histogramr}
\alias{plot.histogramr}
\alias{lines.histogramr}
\title{Plot and Lines Methods for Irregular Histograms}
\usage{
\method{plot}{histogramr}(x, ...)

\method{lines}{histogramr}(x, ...)
}
\arguments{
\item{x}{a \code{histogramr} object.}

\item{...}{parameters passed to \code{plot}, \code{lines}, or \code{points}.}
}
\value{
An invisible copy of \code{x}.
}
\description{
The \code{plot} and \code{lines} methods for \code{histogramr} objects.
}
\examples{
  histogramr(USArrests$Rape/100, breaks = 7)
}
