% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq.R
\name{plot_freq}
\alias{plot_freq}
\title{Histogram and frequency polygon}
\usage{
plot_freq(x, nclass = "Sturges", density = FALSE, cfp = FALSE)
}
\arguments{
\item{x}{a numerical vector.}

\item{nclass}{a character string. Method for calculation of the number of classes from a numerical random variable.}

\item{density}{logical; if TRUE, the histogram graphic is a representation of frequencies, the counts component of the result; if FALSE, probability densities, component density, are plotted (so that the histogram has a total area of one).}

\item{cfp}{logical, indicating if the plot is an histogram and frequency polygon or a cummulative frequency polygon.}
}
\value{
A ggplot2 object.
}
\description{
Creates an histogram and frequency polygon or a cummulative frequency polygon.
}
\examples{
data <- rnorm(100)
plot_freq(data)
plot_freq(data, nclass = "FD", cfp = TRUE)

# Change graphical parameters
plot_freq(data) +
   theme_void()

}
\seealso{
\code{\link{hist}}, \code{\link{geom_histogram}}, \code{\link{geom_freqpoly}}.
}
