% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_wrangler}
\alias{plot_wrangler}
\title{Wrangles arguments for use in the plot, lines and points functions.}
\usage{
plot_wrangler(x, ...)
}
\arguments{
\item{...}{Passed to \code{plot} down the line.}

\item{object}{A \code{histogramr} object.}
}
\description{
Wrangles arguments for use in the plot, lines and points functions.
}
\keyword{internal}
