% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_freq.R
\name{tab_freq}
\alias{tab_freq}
\title{Frequency table}
\usage{
tab_freq(x, nclass = "Sturges", include.lowest = TRUE, right = FALSE)
}
\arguments{
\item{x}{a numerical vector.}

\item{nclass}{a character string. Method for calculation of the number of classes from a numerical random variable.}

\item{include.lowest}{logical, indicating if an ‘x[i]’ equal to the lowest (or highest, for right = FALSE) ‘breaks’ value should be included.}

\item{right}{logical, indicating if the intervals should be closed on the right (and open on the left) or vice versa.}
}
\value{
A tibble (data frame) with five columns. First column contains class intervals, second is frequency '\emph{f}', then relative frequency '\emph{rf}', cummulative frequency '\emph{cf}' and cummulative relative frecuency '\emph{crf}', respectively.
}
\description{
Creates a classical frequency distribution table.
}
\note{
Classes with zero frequency are dropped from table. This is caused by function \code{group_by} from \code{dplyr} package, however a correction for this behavior will be implemented soon. See \url{https://github.com/tidyverse/dplyr/pull/3492}.
}
\examples{
data <- rnorm(100)
tab_freq(data)
tab_freq(data, nclass = "FD")

# Rename columns
tab_freq(data) \%>\%
   rename("Relative frequency" = rl,
          "Frequency" = f)

}
\seealso{
\code{\link{hist}}, \code{\link{cut}}.
}
